+++ src/pages/SYha/Login/Login.scss
@@ -42,12 +42,6 @@
 
         .loginbtn:disabled {
           background-color: #c4e1fb;
-          border-radius: 10px;
-          height: 50px;
-          width: 360px;
-          border: 1px solid white;
-          color: white;
-          font-size: 20px;
         }
         .loginbtn {
           background-color: #0095f6;
