+++ src/pages/SYha/Main/Main.scss
@@ -69,6 +69,7 @@
       max-height: 896px;
       display: flex;
       justify-content: space-around;
+      margin-top: 90px;
     }
   }
