+++ src/pages/SYha/Main/components/Instainfo.js
@@ -1,0 +1,14 @@
+function Instainfo() {
+  return (
+    <div className="instainfo">
+      <p>
+        Instagram 정보 &middot; 지원 &middot; 홍보 &middot; 센터 &middot; API
+        &middot; 채용정보 &middot; 개인정보처리방침 &middot; 약관 &middot;
+        디렉터리 &middot; 프로필 &middot; 해시태그 &middot; 언어
+        <br />
+        <br />© 2019 INSTAGRAM
+      </p>
+    </div>
+  );
+}
+export default Instainfo;
