+++ src/pages/SYha/Main/components/Wecode.js
@@ -1,0 +1,17 @@
+function Wecode() {
+  return (
+    <div className="wecode flex">
+      <img
+        className="wecode_logo img"
+        src="https://velog.velcdn.com/images/inyong_pang/post/f0ea605d-c2d9-460c-aedc-a0ec77e6759f/wecode.png"
+        alt=""
+      />
+      <p>
+        <strong>wecode_bootcamp</strong>
+        <br />
+        <span style={{ color: '#9d9d9d' }}>WeCode | 위코드</span>
+      </p>
+    </div>
+  );
+}
+export default Wecode;
